Example implementation of legacy page converted to controller
